+++ crates/orchestrator/src/jobs/da_job/mod.rs
@@ -64,6 +64,7 @@ pub struct DaJob;
 
 #[async_trait]
 impl Job for DaJob {
+    #[tracing::instrument(fields(category = "da"), skip(self, _config, metadata))]
     async fn create_job(
         &self,
         _config: Arc<Config>,
@@ -83,6 +84,7 @@ impl Job for DaJob {
         })
     }
 
+    #[tracing::instrument(fields(category = "da"), skip(self, config))]
     async fn process_job(&self, config: Arc<Config>, job: &mut JobItem) -> Result<String, JobError> {
         let block_no = job
             .internal_id
@@ -147,6 +149,7 @@ impl Job for DaJob {
         Ok(external_id)
     }
 
+    #[tracing::instrument(fields(category = "da"), skip(self, config))]
     async fn verify_job(&self, config: Arc<Config>, job: &mut JobItem) -> Result<JobVerificationStatus, JobError> {
         Ok(config
             .da_client()
@@ -169,6 +172,7 @@ impl Job for DaJob {
     }
 }
 
+#[tracing::instrument(skip(elements))]
 pub fn fft_transformation(elements: Vec<BigUint>) -> Vec<BigUint> {
     let xs: Vec<BigUint> = (0..*BLOB_LEN)
         .map(|i| {