+++ CHANGELOG.md
@@ -6,7 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## Added
 
-- `SnosJob` implementation
+- `SnosJob` implementation and e2e
+- Telemetry tracing and metrics.
 - e2e flow test
 - database timestamps
 - alerts module.
@@ -53,6 +54,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## Fixed
 
+- Added new_with_settings to SharpClient.
 - Calculate root hash logic and added a simple test for it.
 - Cargo.toml dependency reorg.
 - Get Fact Info logic.